requirements.txt
httpx>=0.25.0
pydantic>=2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
tenacity>=8.2.0
PyYAML>=6.0
rich>=13.0.0